requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-docx==1.1.0
PyMuPDF==1.23.8
reportlab==4.0.8
